modifié :         src/CommonSOG.h 	modifié :         src/HybridKripke.cpp 	modifié :         src/HybridKripkeIterator.cpp 	modifié :         src/HybridKripkeIterator.h 	modifié :         src/HybridKripkeState.cpp 	modifié :         src/HybridKripkeState.h 	modifié :         src/LDDGraph.cpp 	modifié :         src/LDDGraph.h 	modifié :         src/LDDState.cpp 	modifié :         src/LDDState.h 	modifié :         src/MCHybridSOG.cpp 	modifié :         src/MCHybridSOG.h 	modifié :         src/NewNet.cpp 	modifié :         src/NewNet.h 	modifié :         src/SogKripke.cpp 	modifié :         src/SogKripkeTh.cpp 	modifié :         src/main.cpp 	modifié :         src/threadSOG.cpp

+++ src/LDDGraph.cpp
@@ -197,9 +197,9 @@ LDDState *LDDGraph::getLDDStateById(unsigned int id) {
     return m_GONodes.at(id);
 }
 
-string LDDGraph::getTransition(int pos) {
+string LDDGraph::getTransition(uint16_t pos) {
 
-    map<string,int>::iterator it=m_transition->begin();
+    map<string,uint16_t>::iterator it=m_transition->begin();
     while(it != m_transition->end())
     {
         if(it->second == pos)
@@ -209,14 +209,14 @@ string LDDGraph::getTransition(int pos) {
     return it->first;
 }
 
-string LDDGraph::getPlace(int pos) {
+string LDDGraph::getPlace(uint16_t pos) {
     return m_places->find(pos)->second;
 }
 
-void LDDGraph::setTransition(map<string,int>& list_transitions) {
+void LDDGraph::setTransition(map<string,uint16_t>& list_transitions) {
     m_transition=&list_transitions;
 }
-void LDDGraph::setPlace(map<int,string>& list_places) {
+void LDDGraph::setPlace(map<uint16_t,string>& list_places) {
     m_places=&list_places;
 }
